+++ common/embed/credit.d.ts
@@ -1,0 +1,5 @@
+export function createFlourishCredit(credit_url: any, query_string: any, public_url: any, credit_text: any): any;
+export function getLocalizedCreditTextAndUrl(lang: any, credit_key: any): {
+    credit_text: any;
+    credit_url: any;
+};