+++ common/embed/parse_query_params.d.ts
@@ -1,0 +1,2 @@
+export default parseQueryParams;
+declare function parseQueryParams(): {};