+++ common/utils/json.d.ts
@@ -1,0 +1,3 @@
+export function safeStringify(obj: any): string | undefined;
+export function javaScriptStringify(v: any): any;
+export function stringifyPreparedData(data: any): string;