+++ server/comms_js.js
@@ -10,6 +10,7 @@ const CHECK_ORIGIN = `
 		|| (a.hostname.match(/\\.flourish\\.local$/) && window.location.hostname.match(/\\.flourish\\.local$/))
 		|| (a.hostname.match(/\\.flourish\\.net$/) && window.location.hostname.match(/\\.flourish\\.net$/))
 		|| (a.hostname.match(/\\.flourish\\.rocks$/) && window.location.hostname.match(/\\.flourish\\.rocks$/))
+		|| (a.hostname.match(/\\.flourish-internal\\.com$/) && window.location.hostname.match(/\\.flourish-internal\\.com$/))
 		|| (a.hostname.match(/\\.flourish\\.studio$/) && window.location.hostname.match(/\\.flourish\\.studio$/))
 		|| (a.hostname == "app.flourish.studio" && window.location.hostname == "flourish-user-templates.com")
 		|| (a.hostname == "app.flourish.studio" && window.location.hostname == "flourish-user-preview.com")